[REFACTOR]: Rename match_request_service to match_service and update references; add category_id to JobAdPreview schema.
